refactoring

+++ app/controllers/assignments_controller.rb
@@ -1,4 +1,6 @@
 class AssignmentsController < ApplicationController
+    before_action :authenticate_user!
+
     def create
         if !Assignment.find_by(assignment_params)
             @assignment = Assignment.new(assignment_params)
@@ -9,16 +11,21 @@ class AssignmentsController < ApplicationController
             end
         else
             @assignment.errors.add(assignment: 'Assignment Already Exists')
-            render json: {message: @assignment.errors.full_messages}, status: 500
+            render json: {message: @assignment.errors.full_messages[0]}, status: 500
         end
     end
 
     def destroy
         @assignment = Assignment.find_by(assignment_params)
-        if @assignment.destroy
-            render json: { message: 'Assignment successfully destroyed'}, status: 200
+        if current_user == @assignment.detail.creator ||
+            current_user == @assignment.detail.task.project.team.leader
+            if @assignment.destroy 
+                render json: { message: 'Assignment successfully destroyed'}, status: 200
+            else
+                render json: { message: 'Something Went Wrong!'}, status: 500
+            end
         else
-            render json: { message: 'Something Went Wrong!'}, status: 500
+            render json: { message: 'You are not authorized to do that.'}, status: 500
         end
     end
